move tests to the new/agreed location
